requirements.txt
aiogram==3.19.0
python-dotenv==1.1.0